[ #4743 ] Replaced envModality, removed some instances.
The field envModality was removed in favour of envRelevance and
envQuantity, because (according to some comment) cohesion shouldn't be
in the environment. The LensModality instance for TCEnv was removed.

The LensRelevance and LensQuantity instances for TCEnv were also
removed. Now every change of the environment's relevance and quantity
components use the lenses eRelevance and eQuantity.