[ #4743, #6271 ] Added a hard compile-time mode.
When the hard compile-time mode is used all constants that are added
to the signature are made erased. The hard compile-time mode is
entered when an erased definition is checked, but not when a
type-signature is checked.

This is related to issue #6271: When an erased constructor is
type-checked the hard compile-time mode is entered, and all associated
definitions become erased.